Add files via upload
I did the stuff i needed to do along with a asset/description list of the sprites i also did a chicken run cycle and a chicken stop animation i'm unsure if i want to export them or keep them as the files so for now i'm not committing them because for those sprites i used another pixel art program that isn't piskel and you can't open the files without having the program but i'm still unsure because i may want to edit that sprite in my program. Also on another not the chicken sounds i will commit tomorrow i wasn't able to go to my friends today because i was busy so wait on that too, anyways here ya go.